Seperate the player logic into its own script
